refactor: [#1189] extract mod whitelist::manager
